Add files via upload
so this v1, in this script you're going to use the "file path" component in grasshopper to select the .fspy file which you get through the fspy software after adjusting your parameters and saving the file (ctrl + shift + s). The file is opened as text inside of grasshopper, in the example i tried out, out of more than 27,000 lines of code only the first one is used which contains a json and a bunch of giberish characters before and after the json, the grasshopper script is mainly trying to clean that line of code into a pure json and then with python component the json is read and the the data needed is mapped to outputs.

problem #1; what do we do with the data and how do we translate that to rhino? 
A/ https://www.youtube.com/watch?v=gFXbeDetqpU&t=423s 
I actually don't know this youtuber, but i'll try to contact him to collaborate on this project, however i dont actually know how accurate is his script. 

problem #2; some of the data available in the fspy software is not available in the json that we get?
A/ yea, idk this one either, however blender does get everything from that same .fspy file so maybe the data thats missing is somehow the product of calculating some of the other data that is available in the json. as I said I rly dont know and i think issa fun project to put some time into.